chore(InfosPane): store editor top in ISLSError and use it.
We save the editor top in ISLSError and use it to semi-dynamically
calculate margin-tops for the the error messages displayed in the
InfosPane.
Semi-dynamically because their is still some hard coded offset in
the calculation present.
After this commit at least errors should show up in correct place
for multiple vertically stacked editors.